+++ src/Utils.php
@@ -481,6 +481,9 @@ final class Utils
      *   inputs are always treated as string bodies, even when they name callable
      *   functions.
      *
+     * Passing a non-string scalar (`int`, `float`, or `bool`) is deprecated; cast
+     * it to a string instead. guzzlehttp/psr7 3.0 will reject non-string scalars.
+     *
      * @param resource|string|int|float|bool|StreamInterface|callable|\Iterator|\Stringable|null $resource Entity body data
      * @param array{size?: int, metadata?: array}                                                $options  Additional options
      *
@@ -489,8 +492,17 @@ final class Utils
     public static function streamFor($resource = '', array $options = []): StreamInterface
     {
         if (is_scalar($resource)) {
-            if (is_float($resource) && !is_finite($resource)) {
-                throw new \InvalidArgumentException('Cannot create a stream from a non-finite float.');
+            if (!is_string($resource)) {
+                \trigger_deprecation(
+                    'guzzlehttp/psr7',
+                    '2.12',
+                    'Passing %s to Utils::streamFor() is deprecated; cast it to a string. guzzlehttp/psr7 3.0 will only accept string, resource, StreamInterface, Stringable, Iterator, callable, or null.',
+                    \gettype($resource)
+                );
+
+                if (is_float($resource) && !is_finite($resource)) {
+                    throw new \InvalidArgumentException('Cannot create a stream from a non-finite float.');
+                }
             }
 
             $stream = self::tryFopen('php://temp', 'r+');